Add offline version of PRISM
Use _get_weights_path method to retrieve an optional
weights_path given during init or look at the path in the model
registry.
If a valid path is found, the load the offline model from there.
Otherwise, ensure that there is an internet access, then use the
online version of PRISM.
Also removed a trailing whitespace in the forward method.